requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
attrs==18.1.0
autobahn==18.5.2
Automat==0.6.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3
dateparser==0.7.0
future==0.17.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.3.3
mccabe==0.6.1
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
pylint==2.2.2
pyOpenSSL==18.0.0
python-binance==0.6.9
python-dateutil==2.7.3
python-telegram-bot==11.1.0
pytz==2018.4
regex==2018.6.6
requests==2.20.0
service-identity==17.0.0
six==1.11.0
soupsieve==1.9
Twisted==18.4.0
txaio==2.10.0
tzlocal==1.5.1
urllib3==1.23
wrapt==1.10.11
zope.interface==4.5.0
